split test files, seems not more readable :(
