+++ core/core-api/src/test/groovy/com/hbc/pms/core/api/TestDataFixture.groovy
@@ -2,15 +2,15 @@ package com.hbc.pms.core.api
 
 import com.hbc.pms.core.api.controller.v1.request.CreateAlarmConditionCommand
 import com.hbc.pms.core.api.controller.v1.request.UpdateAlarmConditionCommand
-import com.hbc.pms.core.api.service.AlarmConditionPersistenceService
-import com.hbc.pms.core.api.service.BlueprintPersistenceService
-import com.hbc.pms.core.api.service.SensorConfigurationPersistenceService
 import com.hbc.pms.core.api.utils.StringUtils
 import com.hbc.pms.core.model.enums.AlarmActionType
 import com.hbc.pms.core.model.enums.AlarmSeverity
+import com.hbc.pms.core.model.enums.AlarmStatus
 import com.hbc.pms.core.model.enums.AlarmType
 import com.hbc.pms.core.model.enums.BlueprintType
+import com.hbc.pms.integration.db.entity.AlarmActionEntity
 import com.hbc.pms.integration.db.entity.AlarmConditionEntity
+import com.hbc.pms.integration.db.entity.AlarmHistoryEntity
 import com.hbc.pms.integration.db.entity.BlueprintEntity
 import com.hbc.pms.integration.db.entity.SensorConfigurationEntity
 import com.hbc.pms.integration.db.repository.AlarmActionRepository
@@ -18,7 +18,7 @@ import com.hbc.pms.integration.db.repository.AlarmConditionRepository
 import com.hbc.pms.integration.db.repository.AlarmHistoryRepository
 import com.hbc.pms.integration.db.repository.BlueprintRepository
 import com.hbc.pms.integration.db.repository.SensorConfigurationRepository
-import com.hbc.pms.plc.api.PlcConnector
+import java.time.OffsetDateTime
 import java.util.concurrent.ThreadLocalRandom
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
@@ -33,18 +33,6 @@ class TestDataFixture {
   static Long MONITORING_BLUEPRINT_ID
   static Long CUSTOM_ALARM_BLUEPRINT_ID
   static Long PREDEFINED_ALARM_BLUEPRINT_ID
-//  static Long CUSTOM_ALARM_CONDITION_ID
-//  static Long PREDEFINED_ALARM_CONDITION_ID
-  static Long REAL_SENSOR_WITH_CONDITION_ID
-  static Long REAL_SENSOR_WITHOUT_CONDITION_ID
-//  static Long BOOL_SENSOR_WITH_CONDITION_ID
-//  static Long BOOL_SENSOR_WITHOUT_CONDITION_ID
-//  static Long HISTORY_WITH_EMAIL_ACTION_ID
-//  static Long HISTORY_WITH_POPUP_ACTION_ID
-//  static Long HISTORY_WITH_TWO_ACTIONS_ID
-
-  @Autowired
-  AlarmConditionPersistenceService alarmConditionPersistenceService
 
   @Autowired
   BlueprintRepository blueprintRepository
@@ -61,15 +49,6 @@ class TestDataFixture {
   @Autowired
   SensorConfigurationRepository configurationRepository
 
-  @Autowired
-  BlueprintPersistenceService blueprintPersistenceService
-
-  @Autowired
-  SensorConfigurationPersistenceService configurationPersistenceService
-
-  @Autowired
-  PlcConnector connector
-
   void populateDefaultBlueprints() {
     def monitoringBlueprint = blueprintRepository.save(createBlueprint(BlueprintType.MONITORING, "Monitoring"))
     MONITORING_BLUEPRINT_ID = monitoringBlueprint.getId()
@@ -79,35 +58,6 @@ class TestDataFixture {
 
     def customAlarmBlueprint = blueprintRepository.save(createBlueprint(BlueprintType.ALARM, AlarmType.CUSTOM.toString()))
     CUSTOM_ALARM_BLUEPRINT_ID = customAlarmBlueprint.getId()
-
-//    def realMonitoringConfig
-//            = configurationRepository.save(createSensorConfiguration(monitoringBlueprint, PLC_ADDRESS_REAL_01))
-//    REAL_SENSOR_WITH_CONDITION_ID = realMonitoringConfig.getId()
-//
-//    def realMonitoringConfig2 = configurationRepository.save(createSensorConfiguration(monitoringBlueprint, PLC_ADDRESS_REAL_02))
-//    REAL_SENSOR_WITHOUT_CONDITION_ID = realMonitoringConfig2.getId()
-//
-//    def boolMonitoringConfig
-//            = configurationRepository.save(createSensorConfiguration(monitoringBlueprint, PLC_ADDRESS_BOOL_01))
-//    BOOL_SENSOR_WITH_CONDITION_ID = boolMonitoringConfig.getId()
-//
-//    def customAlarmCondition = alarmConditionRepository.save(createDefaultConditionEntity(realMonitoringConfig))
-//
-//    var sensor = configurationRepository.save(createSensorConfiguration(createdBlueprint.id, PLC_ADDRESS_REAL_01))
-//    configurationRepository.save(createSensorConfiguration(createdBlueprint.id, PLC_ADDRESS_REAL_02))
-//    var alarmCondition = alarmConditionRepository.save(createDefaultConditionEntity(sensor))
-//    alarmActionRepository.save(createAction(alarmCondition))
-//    alarmHistoryRepository.save(createHistory(alarmCondition))
-
-    connector.updateScheduler()
-  }
-
-  void populateSensorConfigs() {
-    def monitoringBlueprint = blueprintRepository.findById(MONITORING_BLUEPRINT_ID).get()
-
-    def realMonitoringConfig
-            = configurationRepository.save(createSensorConfiguration(monitoringBlueprint, PLC_ADDRESS_REAL_01))
-    REAL_SENSOR_WITH_CONDITION_ID = realMonitoringConfig.getId()
   }
 
   void cleanup() {
@@ -133,22 +83,22 @@ class TestDataFixture {
             .build()
   }
 
-//  static AlarmHistoryEntity createHistory(AlarmConditionEntity alarmCondition) {
-//    return AlarmHistoryEntity.builder()
-//            .status(AlarmStatus.SOLVED)
-//            .triggeredAt(OffsetDateTime.now())
-//            .condition(alarmCondition)
-//            .build()
-//  }
-//
-//  static AlarmActionEntity createAction(AlarmConditionEntity alarmCondition) {
-//    return AlarmActionEntity.builder()
-//            .condition(alarmCondition)
-//            .type(AlarmActionType.POPUP)
-//            .message("Popup Action message")
-//            .build()
-//  }
-//
+  static AlarmHistoryEntity createHistory(AlarmConditionEntity alarmCondition) {
+    return AlarmHistoryEntity.builder()
+            .status(AlarmStatus.SOLVED)
+            .triggeredAt(OffsetDateTime.now())
+            .condition(alarmCondition)
+            .build()
+  }
+
+  static AlarmActionEntity createAction(AlarmConditionEntity alarmCondition) {
+    return AlarmActionEntity.builder()
+            .condition(alarmCondition)
+            .type(AlarmActionType.POPUP)
+            .message("Popup Action message")
+            .build()
+  }
+
   static AlarmConditionEntity createDefaultConditionEntity(AlarmType alarmType, SensorConfigurationEntity sensorConfig) {
     return AlarmConditionEntity.builder()
             .isEnabled(false)
