+++ core/core-api/src/test/groovy/com/hbc/pms/core/api/controllers/AlarmConditionControllerCreateEndpointsFunctionalSpec.groovy
@@ -4,44 +4,88 @@ import com.hbc.pms.core.api.TestDataFixture
 import com.hbc.pms.core.api.support.error.ErrorCode
 import com.hbc.pms.core.api.support.response.ApiResponse
 import com.hbc.pms.core.api.test.setup.FunctionalTestSpec
-import spock.lang.PendingFeature
+import com.hbc.pms.core.model.enums.AlarmType
+import com.hbc.pms.integration.db.repository.AlarmActionRepository
+import com.hbc.pms.integration.db.repository.AlarmConditionRepository
+import com.hbc.pms.integration.db.repository.BlueprintRepository
+import com.hbc.pms.integration.db.repository.SensorConfigurationRepository
+import com.hbc.pms.support.spock.test.RestClient
+import org.springframework.beans.factory.annotation.Autowired
 
 class AlarmConditionControllerCreateEndpointsFunctionalSpec extends FunctionalTestSpec {
-  def setup() {
+  @Autowired
+  BlueprintRepository blueprintRepository
 
+  @Autowired
+  SensorConfigurationRepository configurationRepository
+
+  @Autowired
+  AlarmConditionRepository conditionRepository
+
+  @Autowired
+  AlarmActionRepository actionRepository
+
+  @Autowired
+  RestClient restClient
+
+  Long REAL_SENSOR_WITHOUT_CONDITION_ID
+  Long REAL_SENSOR_WITH_CONDITION_ID
+
+  def setup() {
+    def monitorBlueprint
+            = blueprintRepository.findById(TestDataFixture.MONITORING_BLUEPRINT_ID).get()
+    def sensorConfig
+            = configurationRepository
+            .save(
+                    TestDataFixture.createSensorConfiguration(monitorBlueprint, TestDataFixture.PLC_ADDRESS_REAL_01)
+            )
+    REAL_SENSOR_WITHOUT_CONDITION_ID = sensorConfig.getId()
+
+    def sensorConfig2
+            = configurationRepository
+            .save(
+                    TestDataFixture.createSensorConfiguration(monitorBlueprint, TestDataFixture.PLC_ADDRESS_BOOL_02)
+            )
+    REAL_SENSOR_WITH_CONDITION_ID = sensorConfig2.getId()
+    conditionRepository.save(TestDataFixture.createDefaultConditionEntity(AlarmType.PREDEFINED, sensorConfig2))
   }
 
   def "Create new alarm condition - OK"() {
     given:
     def sensorConfiguration
-            = sensorConfigurationRepository.findById(TestDataFixture.REAL_SENSOR_WITHOUT_CONDITION_ID).get()
-    def createConditionCommand = createDefaultAlarmConditionCommand(sensorConfiguration)
+            = configurationRepository.findById(REAL_SENSOR_WITHOUT_CONDITION_ID).get()
+    def createConditionCommand
+            = TestDataFixture.createDefaultAlarmConditionCommand(sensorConfiguration)
 
     when:
-    def response = restClient.post("/alarm-conditions", createConditionCommand, ApiResponse.class)
+    def response
+            = restClient.post("/alarm-conditions", createConditionCommand, ApiResponse.class)
 
     then:
     response.statusCode.is2xxSuccessful()
     int createdConditionId = response.body.data["id"] as int
-    def createdCondition = alarmConditionRepository.findById(createdConditionId)
+    def createdCondition = conditionRepository.findById(createdConditionId)
     assert createdCondition.isPresent()
     assert createdCondition.get().getSensorConfiguration().id == sensorConfiguration.id
   }
 
   def "Create new alarm condition - Invalid value (#value) for checkInterval - Bad request"() {
     given:
-    def sensorConfiguration = sensorConfigurationRepository.findAllByAddress(TestDataFixture.PLC_ADDRESS_REAL_02).first()
-    def createConditionCommand = createDefaultAlarmConditionCommand(sensorConfiguration)
+    def sensorConfiguration
+            = configurationRepository.findById(REAL_SENSOR_WITHOUT_CONDITION_ID).get()
+    def createConditionCommand
+            = TestDataFixture.createDefaultAlarmConditionCommand(sensorConfiguration)
     createConditionCommand.setCheckInterval(value)
-    def conditionCountBefore = alarmConditionRepository.findAll().size()
+    def conditionCountBefore = conditionRepository.findAll().size()
 
     when:
-    def response = restClient.post("/alarm-conditions", createConditionCommand, ApiResponse.class)
+    def response
+            = restClient.post("/alarm-conditions", createConditionCommand, ApiResponse.class)
 
     then:
     response.statusCode.is4xxClientError()
     response.body.error["code"] == ErrorCode.E400.toString()
-    def conditionCountAfter = alarmConditionRepository.findAll().size()
+    def conditionCountAfter = conditionRepository.findAll().size()
     conditionCountBefore == conditionCountAfter
 
     where:
@@ -50,19 +94,22 @@ class AlarmConditionControllerCreateEndpointsFunctionalSpec extends FunctionalTe
 
   def "Create new alarm condition - Invalid value (#value) for timeDelay - Bad request"() {
     given:
-    def sensorConfiguration = sensorConfigurationRepository.findAllByAddress(TestDataFixture.PLC_ADDRESS_REAL_02).first()
-    def createConditionCommand = createDefaultAlarmConditionCommand(sensorConfiguration)
+    def sensorConfiguration
+            = configurationRepository.findById(REAL_SENSOR_WITHOUT_CONDITION_ID).get()
+    def createConditionCommand
+            = TestDataFixture.createDefaultAlarmConditionCommand(sensorConfiguration)
     createConditionCommand.setTimeDelay(value)
-    def conditionCountBefore = alarmConditionRepository.findAll().size()
+    def conditionCountBefore = conditionRepository.findAll().size()
 
     when:
-    def response = restClient.post("/alarm-conditions", createConditionCommand, ApiResponse.class)
+    def response
+            = restClient.post("/alarm-conditions", createConditionCommand, ApiResponse.class)
 
     then:
     response.statusCode.is4xxClientError()
     response.body.error["code"] == ErrorCode.E400.toString()
     response.body.error.data["timeDelay"]
-    def conditionCountAfter = alarmConditionRepository.findAll().size()
+    def conditionCountAfter = conditionRepository.findAll().size()
     conditionCountBefore == conditionCountAfter
 
     where:
@@ -72,67 +119,72 @@ class AlarmConditionControllerCreateEndpointsFunctionalSpec extends FunctionalTe
   def "Create new alarm condition - CUSTOM condition with both min & max = null - Bad request"() {
     given:
     def sensorConfiguration
-            = sensorConfigurationRepository.findAllByAddress(TestDataFixture.PLC_ADDRESS_REAL_02).first()
-    def createConditionCommand = createDefaultAlarmConditionCommand(sensorConfiguration)
+            = configurationRepository.findAllByAddress(TestDataFixture.PLC_ADDRESS_REAL_01).first()
+    def createConditionCommand
+            = TestDataFixture.createDefaultAlarmConditionCommand(sensorConfiguration)
     createConditionCommand.setMin(null)
     createConditionCommand.setMax(null)
-    def conditionCountBefore = alarmConditionRepository.findAll().size()
+    def conditionCountBefore = conditionRepository.findAll().size()
 
     when:
-    def response = restClient.post("/alarm-conditions", createConditionCommand, ApiResponse.class)
+    def response
+            = restClient.post("/alarm-conditions", createConditionCommand, ApiResponse.class)
 
     then:
     response.statusCode.is4xxClientError()
     response.body.error["code"] == ErrorCode.E400.toString()
-    def conditionCountAfter = alarmConditionRepository.findAll().size()
+    def conditionCountAfter = conditionRepository.findAll().size()
     conditionCountBefore == conditionCountAfter
 
     where:
     value << [-1, 0, 3601]
   }
 
-  @PendingFeature
   def "Create new alarm condition - CUSTOM condition with min greater than or equal max - Bad request"() {
     given:
     def sensorConfiguration
-            = sensorConfigurationRepository.findAllByAddress(TestDataFixture.PLC_ADDRESS_REAL_02).first()
-    def createConditionCommand = createDefaultAlarmConditionCommand(sensorConfiguration)
+            = configurationRepository.findById(REAL_SENSOR_WITH_CONDITION_ID).get()
+    def createConditionCommand
+            = TestDataFixture.createDefaultAlarmConditionCommand(sensorConfiguration)
     createConditionCommand.setMin(120)
     createConditionCommand.setMax(12)
-    def conditionCountBefore = alarmConditionRepository.findAll().size()
+    def conditionCountBefore = conditionRepository.findAll().size()
 
     when:
-    def response = restClient.post("/alarm-conditions", createConditionCommand, ApiResponse.class)
+    def response
+            = restClient.post("/alarm-conditions", createConditionCommand, ApiResponse.class)
 
     then:
     response.statusCode.is4xxClientError()
     response.body.error["code"] == ErrorCode.E400.toString()
-    def conditionCountAfter = alarmConditionRepository.findAll().size()
+    def conditionCountAfter = conditionRepository.findAll().size()
     conditionCountBefore == conditionCountAfter
   }
 
   def "Create new alarm condition - Existed condition with Tag configuration - Bad request"() {
     given:
     def sensorConfiguration
-            = sensorConfigurationRepository.findAllByAddress(TestDataFixture.PLC_ADDRESS_REAL_01).first()
-    def createConditionCommand = createDefaultAlarmConditionCommand(sensorConfiguration)
-    def conditionCountBefore = alarmConditionRepository.findAll().size()
+            = configurationRepository.findById(REAL_SENSOR_WITH_CONDITION_ID).get()
+    def createConditionCommand
+            = TestDataFixture.createDefaultAlarmConditionCommand(sensorConfiguration)
+    def conditionCountBefore = conditionRepository.findAll().size()
 
     when:
-    def response = restClient.post("/alarm-conditions", createConditionCommand, ApiResponse.class)
+    def response
+            = restClient.post("/alarm-conditions", createConditionCommand, ApiResponse.class)
 
     then:
     response.statusCode.is4xxClientError()
     response.body.error["code"] == ErrorCode.E400.toString()
     response.body.error["data"].containsIgnoreCase("Existed")
-    def conditionCountAfter = alarmConditionRepository.findAll().size()
+    def conditionCountAfter = conditionRepository.findAll().size()
     conditionCountBefore == conditionCountAfter
   }
 
   def "Create new alarm action - OK"() {
     given:
-    def condition = alarmConditionRepository.findAll().first()
-    def createActionCommand = createDefaultAlarmActionCommand()
+    def condition = conditionRepository.findAll().first()
+    def createActionCommand = TestDataFixture.createDefaultAlarmActionCommand()
 
     when:
     def response
@@ -143,15 +195,15 @@ class AlarmConditionControllerCreateEndpointsFunctionalSpec extends FunctionalTe
     then:
     response.statusCode.is2xxSuccessful()
     int createdActionId = response.body.data["id"] as int
-    def createdAction = alarmActionRepository.findById(createdActionId)
+    def createdAction = actionRepository.findById(createdActionId)
     createdAction.isPresent()
     createdAction.get().getCondition().getId() == condition.getId()
   }
 
   def "Create new alarm action - Not existing alarm condition - Not found and Bad request"() {
     given:
-    def createActionCommand = createDefaultAlarmActionCommand()
-    def actionCountBefore = alarmActionRepository.findAll().size()
+    def createActionCommand = TestDataFixture.createDefaultAlarmActionCommand()
+    def actionCountBefore = actionRepository.findAll().size()
 
     when:
     def response
@@ -163,16 +215,16 @@ class AlarmConditionControllerCreateEndpointsFunctionalSpec extends FunctionalTe
     response.statusCode.is4xxClientError()
     response.body.error["code"] == ErrorCode.E404.toString()
     response.body.error["data"].containsIgnoreCase("Not found")
-    def actionCountAfter = alarmActionRepository.findAll().size()
+    def actionCountAfter = actionRepository.findAll().size()
     actionCountBefore == actionCountAfter
   }
 
   def "Create new alarm action - EMAIL action with no recipient - Bad request"() {
     given:
-    def condition = alarmConditionRepository.findAll().first()
-    def createActionCommand = createDefaultAlarmActionCommand()
+    def condition = conditionRepository.findAll().first()
+    def createActionCommand = TestDataFixture.createDefaultAlarmActionCommand()
     createActionCommand.setRecipients(null)
-    def actionCountBefore = alarmActionRepository.findAll().size()
+    def actionCountBefore = actionRepository.findAll().size()
 
     when:
     def response
@@ -184,7 +236,7 @@ class AlarmConditionControllerCreateEndpointsFunctionalSpec extends FunctionalTe
     response.statusCode.is4xxClientError()
     response.body.error["code"] == ErrorCode.E400.toString()
     response.body.error.data["validRecipients"]
-    def actionCountAfter = alarmActionRepository.findAll().size()
+    def actionCountAfter = actionRepository.findAll().size()
     actionCountBefore == actionCountAfter
   }
 }
